backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pandas>=2.2.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.25
aiofiles>=23.2.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
